Fixing review comment - Modified Integer to Long in Repo

+++ src/main/java/com/musicshare/repository/PlaylistRepository.java
@@ -3,5 +3,5 @@ package com.musicshare.repository;
 import com.musicshare.entity.Playlist;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PlaylistRepository extends JpaRepository<Playlist, Integer> {
+public interface PlaylistRepository extends JpaRepository<Playlist, Long> {
 }
